+++ PandorasBox/Helpers/NumberHelper.cs
@@ -7,7 +7,7 @@ namespace PandorasBox.Helpers
         public static int RoundOff(this int i, int sliderIncrement)
         {
             var sliderAsDouble = Convert.ToDouble(sliderIncrement);
-            return ((int)Math.Round(i / sliderAsDouble)) * (int)sliderIncrement;
+            return ((int)Math.Round(i / sliderAsDouble)) * sliderIncrement;
         }
 
         public static float RoundOff(this float i, float sliderIncrement)
